+++ src/connectedk8s/azext_connectedk8s/_utils.py
@@ -238,18 +238,24 @@ def check_cluster_outbound_connectivity(outbound_connectivity_check_log, filepat
         outbound_connectivity_response = outbound_connectivity_response[::-1]
         if consts.Outbound_Connectivity_Check_Result_String not in outbound_connectivity_check_log:
             return consts.Diagnostic_Check_Incomplete, storage_space_available
+
+        # extracting the obo endpoint log specifically from outbound connectivity check job logs
+        obo_check_log = outbound_connectivity_check_log.split('  ')[0]
+        # extracting the obo endpoint url
+        obo_enpoint_url = obo_check_log.split(" : ")[1]
+        # extracting the obo endpoint response code from outbound connectivity check
+        obo_check_response_code = obo_check_log.split(" : ")[2]
         # Validating if outbound connectiivty is working or not and displaying proper result
         if(outbound_connectivity_response != "000"):
-            if consts.Outbound_Connectivity_Failed_String_For_OBO_Endpoint in outbound_connectivity_check_log:
-                failed_obo_enpoint = outbound_connectivity_check_log.split(consts.Outbound_Connectivity_Check_Result_String)[0]
-                failed_obo_enpoint_url = failed_obo_enpoint.split(consts.Outbound_Connectivity_Failed_String_For_OBO_Endpoint)[1]
-                logger.warning("The outbound network connectivity check for has failed for the endpoint - " + failed_obo_enpoint_url + "\nThis will affect the \"cluser-connect\" functionality. If you have configured an outbound proxy in your environment, please ensure that all endpoints are properly whitelisted by referring to the network requirements outlined here: https://learn.microsoft.com/en-us/azure/azure-arc/network-requirements-consolidated?tabs=azure-cloud#azure-arc-enabled-kubernetes-endpoints.\n")
-                telemetry.set_exception(exception='Outbound network connectivity check failed for the OBO endpoint', fault_type=consts.Outbound_Connectivity_Check_Failed, summary="Outbound network connectivity check failed for the OBO endpoint")
+            # since connectivity to obo endpoint is required only for cluster connect feature , we only give out a warning if connectivity to obo endpoint is not present
+            if obo_check_response_code == "000":
+                logger.warning("The outbound network connectivity check for has failed for the endpoint - " + obo_enpoint_url + "\nThis will affect the \"cluser-connect\" functionality. If you have configured an outbound proxy in your environment, please ensure that all endpoints are properly whitelisted by referring to the network requirements outlined here: https://learn.microsoft.com/en-us/azure/azure-arc/network-requirements-consolidated?tabs=azure-cloud#azure-arc-enabled-kubernetes-endpoints.\n")
+                telemetry.set_exception(exception='Outbound network connectivity check failed for the OBO endpoint', fault_type=consts.Outbound_Connectivity_Check_Failed_For_OBO, summary="Outbound network connectivity check failed for the OBO endpoint")
             if storage_space_available:
                 outbound_connectivity_check_path = os.path.join(filepath_with_timestamp, consts.Outbound_Network_Connectivity_Check)
                 with open(outbound_connectivity_check_path, 'w+') as outbound:
-                    if consts.Outbound_Connectivity_Failed_String_For_OBO_Endpoint in outbound_connectivity_check_log:
-                        outbound.write("Response code " + outbound_connectivity_response + "\nOutbound network connectivity check passed successfully." + "\nOutbound connectivity failed for the endpoint-" + failed_obo_enpoint_url + "This is an optional endpoint needed for cluster-connect feature.")
+                    if obo_check_response_code == "000":
+                        outbound.write("Response code " + outbound_connectivity_response + "\nOutbound network connectivity check passed successfully." + "\nOutbound connectivity failed for the endpoint:" + obo_enpoint_url + " ,this is an optional endpoint needed for cluster-connect feature.")
                     else:
                         outbound.write("Response code " + outbound_connectivity_response + "\nOutbound network connectivity check passed successfully.")
             return consts.Diagnostic_Check_Passed, storage_space_available